Add support for some disk subroutines to HLE BIOS
Doki Doki Panic, Deep Dungeon II and possibly other games use some of
the file-level I/O subroutines normally used by LoadFiles and friends.
It appears this is done as some sort of anti-piracy measure, at least in
the case of Deep Dungeon II.

Also, the auto-disk-change hack has been improved.  It is based on
6502 clock cycles instead of a simple counter.  Eggerland (Japan's
"Adventures of Lolo") checks for disk presence approximately once per
second instead of checking a certain number of times, so the previous
hack didn't work for it.  This new one should work for every game,
provided that 2 million cycles is enough time between toggles of the
'disk present' bit.  So far it seems to work, but I'll increase the
toggle delay as needed.

Issues:

- I can't guarantee that these calls behave properly for other games
  that might use them.  In particular, the way we tell which file to
  compare in FileMatchTest and load in FileLoad assumes that the routine
  calling these subroutines uses $0006 to hold the current file count.
  This is what LoadFiles does, and the games that use these routines
  directly appear to do so too.  However, there's no reason that a
  different location couldn't be used by calling GetFileCount and
  storing the result elsewhere before the loop.
- Save states should work fine, since the only additional state (which
  file to load/match) is figured out dynamically based on the contents
  of memory at $0006.
- Currently CheckFileType always succeeds, since it's only used to make
  sure that we're looking at a file header block when we expect to be.
  This isn't an issue in HLE mode, so we just make it succeed.