Fix for popping SDL sound
Apparently a "fix" was committed for this a couple of years ago.  Unfortunately
the "fix" was broken too.  It's OK for the write index/pointer of a ring buffer
to be ahead of the read index/pointer.  All the previous change ensured is that
there *would* be "garbage" data sent to the sound card.  The correct fix is to
use SDL_LockAudio() and SDL_UnlockAudio() in WriteSound() to make sure that
s_BufferIn isn't modified by both threads at the same time.