Well, time to start over (sorta)
Since starting this little project I've found a few compatibility issues
with unlicensed games.  Simply put, it looks like they depend on the PPU
being in a certain state after loading files, and my current code
doesn't (and can't) try to replicate that state.  Additionally, the code
is becoming much more complex than I feel it should need to be, and
there are a few nasty hacks (e.g. the NMI hack) that are only necessary
because of the current design.

Instead of continuing along this path, I've decided to scrap most of
what I've done in favor of a somewhat lower-level approach.  The idea is
to trap calls to the XferByte routine and implement the reading and
writing of bytes in the emulator itself.  This means that the disk seek
interrupt doesn't need to be implemented (or managed).  Other addresses
will be trapped as necessary to make this work.  Additionally, some
traps may be added to improve performance.

The win here is maximum compatibility (aside from complete low-level
emulation, of course).  Additionally, there should be a lot less code
and what is there should be pretty straightforward.  As long as a game
doesn't try to read directly from the hardware or make assumptions about
the state of the disk head between calls to XferByte, it should work
fine.

Here's what has been implemented so far:
- XferByte trap to high-level implementation (read operations only)
- Trap to patch FileMatchTest so it doesn't depend on the disk IRQ
  handler
- Patch to skip FDS "title" screen and copyright screen.  This saves 4-5
  seconds.
- Trap added to LoadData routine to skip files by adjusting the disk
  pointer and relevant RAM locations instead of dummy reads.
- Trap added to LoadData routine to load PRG-RAM files via memcpy()
  instead of loading them byte-by-byte.
- Removed several calls to the delay routine by NOPing them out.  This
  saves just under a second.

I've only tried a handful of games (SMB, SMB2, Zelda I & II, and some
unlicensed titles) and they all work fine, as far as loading goes.
Saving obviously doesn't work yet.  Still, with the above traps and
patches SMB2 loads in just over 1.5 seconds, vs. 12.5 seconds under
low-level emulation.  That's pretty close to my old high-level
implementation.  The main difference is due to CHR-RAM files, which must
be loaded byte-by-byte to ensure that the PPU state is accurate.